algorithm in map

+++ app/src/test/java/com/example/isabelmangan/blueprinttravel/AlgoTest.java
@@ -15,7 +15,7 @@ public class AlgoTest {
         //places open later
         int[] open = {0,700,700,700};
         int[] close = {0,750,960,1200};
-        int[][] walking_time = {{0, 30, 60, 120}, {30, 0, 30, 90}, {60, 30, 0, 60}, {120, 90, 60, 0}};
+        long[][] walking_time = {{0, 30, 60, 120}, {30, 0, 30, 90}, {60, 30, 0, 60}, {120, 90, 60, 0}};
         int[] time_spent = {0, 60, 30, 60};
 
         assertNull(Algo.generateOptimizedRoute(curr,open,close,walking_time,time_spent));
@@ -26,7 +26,7 @@ public class AlgoTest {
         //places are closed/will close
         int[] open = {0,400,400,400};
         int[] close = {0,500,650,700};
-        int[][] walking_time = {{0, 30, 60, 120}, {30, 0, 30, 90}, {60, 30, 0, 60}, {120, 90, 60, 0}};
+        long[][] walking_time = {{0, 30, 60, 120}, {30, 0, 30, 90}, {60, 30, 0, 60}, {120, 90, 60, 0}};
         int[] time_spent = {0, 60, 60, 60};
 
         assertNull(Algo.generateOptimizedRoute(curr,open,close,walking_time,time_spent));
@@ -37,7 +37,7 @@ public class AlgoTest {
         //all places visited
         int[] open = {0,400,400,400};
         int[] close = {0,750,660,1020};
-        int[][] walking_time = {{0, 30, 60, 120}, {30, 0, 30, 90}, {60, 30, 0, 60}, {120, 90, 60, 0}};
+        long[][] walking_time = {{0, 30, 60, 120}, {30, 0, 30, 90}, {60, 30, 0, 60}, {120, 90, 60, 0}};
         int[] time_spent = {0, 60, 60, 60};
 
         ArrayList<Integer> output = new ArrayList<>();
@@ -55,7 +55,7 @@ public class AlgoTest {
         int curr = 540;
         int[] open = {0,400,1020,400};
         int[] close = {0,750,660,1020};
-        int[][] walking_time = {{0, 30, 60, 120}, {30, 0, 30, 90}, {60, 30, 0, 60}, {120, 90, 60, 0}};
+        long[][] walking_time = {{0, 30, 60, 120}, {30, 0, 30, 90}, {60, 30, 0, 60}, {120, 90, 60, 0}};
         int[] time_spent = {0, 60, 60, 60};
 
         ArrayList<Integer> output = new ArrayList<>();
@@ -72,7 +72,7 @@ public class AlgoTest {
         int curr = 540;
         int[] open = {0,400,400,400};
         int[] close = {0,630,660,1020};
-        int[][] walking_time = {{0, 30, 60, 120}, {30, 0, 30, 90}, {60, 30, 0, 60}, {120, 90, 60, 0}};
+        long[][] walking_time = {{0, 30, 60, 120}, {30, 0, 30, 90}, {60, 30, 0, 60}, {120, 90, 60, 0}};
         int[] time_spent = {0, 60, 60, 60};
 
         ArrayList<Integer> output = new ArrayList<>();
